reorg files paths
